Move decorations related classes into platform-base
