+++ macos/src/main/java/com/github/weisj/darklaf/platform/macos/MacOSDecorationsProvider.java
@@ -26,10 +26,10 @@ import java.util.List;
 
 import javax.swing.*;
 
+import com.github.weisj.darklaf.platform.CustomTitlePane;
+import com.github.weisj.darklaf.platform.DecorationsProvider;
 import com.github.weisj.darklaf.platform.SystemInfo;
-import com.github.weisj.darklaf.platform.decorations.CustomTitlePane;
-import com.github.weisj.darklaf.platform.decorations.DecorationsProvider;
-import com.github.weisj.darklaf.platform.decorations.UnsupportedProviderException;
+import com.github.weisj.darklaf.platform.UnsupportedProviderException;
 import com.github.weisj.darklaf.platform.macos.ui.MacOSTitlePane;
 
 public class MacOSDecorationsProvider implements DecorationsProvider {
