+++ platform-base/src/main/java/com/github/weisj/darklaf/platform/DecorationsProvider.java
@@ -18,7 +18,7 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.github.weisj.darklaf.platform.decorations;
+package com.github.weisj.darklaf.platform;
 
 import java.awt.*;
 import java.util.List;
