+++ windows/src/main/java/com/github/weisj/darklaf/platform/windows/WindowsDecorationsProvider.java
@@ -27,10 +27,10 @@ import java.util.List;
 import javax.swing.*;
 import javax.swing.border.Border;
 
+import com.github.weisj.darklaf.platform.CustomTitlePane;
+import com.github.weisj.darklaf.platform.DecorationsProvider;
 import com.github.weisj.darklaf.platform.SystemInfo;
-import com.github.weisj.darklaf.platform.decorations.CustomTitlePane;
-import com.github.weisj.darklaf.platform.decorations.DecorationsProvider;
-import com.github.weisj.darklaf.platform.decorations.UnsupportedProviderException;
+import com.github.weisj.darklaf.platform.UnsupportedProviderException;
 import com.github.weisj.darklaf.platform.windows.ui.WindowsTitlePane;
 
 public class WindowsDecorationsProvider implements DecorationsProvider {
